# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: API for WASAText messaging application
  version: 1.0.0
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Create a new session (login)
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        required: true
        description: User details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: The user was found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadReqError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/username:
    post:
      tags: ["users"]
      description: Update the username by sending the new one
      summary: Update your username
      operationId: setMyUserName
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "204":
          description: Username updated successfully
        "400":
          $ref: "#/components/responses/BadReqError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users:
    get:
      tags: ["users"]
      operationId: getUsers
      description: Get a list of users
      summary: Get a list of users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                description: The object containing the user list
                type: object
                properties:
                  userIdList:
                    $ref: "#/components/schemas/UserIdList"
                required:
                  - userIdList
        "400":
          $ref: "#/components/responses/BadReqError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}:
    get:
      tags: ["users"]
      operationId: getUsersNotInConversation
      description: Get a list of users not already in a conversation
      summary: Get a list of users 
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                description: The object containing the user list
                type: object
                properties:
                  userIdList:
                    $ref: "#/components/schemas/UserIdList"
                required:
                  - userIdList
        "400":
          $ref: "#/components/responses/BadReqError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users/{userId}/photo:
    get:
      tags:
        - "users"
        - "photos"
      operationId: getUserPicture
      summary: Retrieve user picture
      description: Retrieve user picture
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: An Image
          content:
            image/png:
              schema:
                description: photo
                type: string
                format: binary
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /photos/{photoId}/users/{userId}/photo:
    parameters:
      - $ref: "#/components/parameters/photoId"
      - $ref: "#/components/parameters/userId"
    post:
      tags: ["photos", "users"]
      summary: Set user photo
      description: Set user photo
      operationId: setMyPhoto
      responses:
        "200":
          description: The photo was changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /photos/{photoId}/groups/{groupId}/photo:
    parameters:
      - $ref: "#/components/parameters/photoId"
      - $ref: "#/components/parameters/groupId"
    post:
      tags: ["photos", "groups"]
      summary: Set group photo
      description: Set group photo
      operationId: setGroupPhoto
      responses:
        "204":
          description: The photo was changed
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /photos/{photoId}:
    parameters:
      - $ref: "#/components/parameters/photoId"
    get:
      tags: ["photos", "messages"]
      summary: Get the photo of the message
      description: Get the photo of the message
      operationId: getPhoto
      responses:
        "200":
          description: Take the photo
          content:
            image/png:
              schema:
                type: string
                format: binary
                description: Photo
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
...
    Photo:
      type: object
      description: Image converted to png
      properties:
        id:
          $ref: "#/components/schemas/Id"
        path:
          $ref: "#/components/schemas/Path"
        image/png:
          type: string
          format: binary
          description: Image converted to png
        size:
          type: integer
          format: int64
          description: size of the Image
          minimum: 1
          maximum: 20000000
    UserIdList:
      type: array
      description: List of user IDs to include in the group
      minItems: 2
      maxItems: 100
      items:
        $ref: "#/components/schemas/UserId"
...
    Id:
      type: integer
      description: Unique identifier for a user
      maximum: 10000000
      minimum: 0
      format: int64
    UserId:
      type: integer
      description: Unique identifier for a user
      maximum: 10000000
      minimum: 0
      format: int64
    UserName:
      type: string
      description: Username for a user account
      minLength: 1
      maxLength: 20
      pattern: "^[a-zA-Z0-9._\\!-]+$"
      example: "Rob-_!2."
    User:
      type: object
      description: Represents a user in the system
      properties:
        id:
          $ref: "#/components/schemas/UserId"
        name:
          $ref: "#/components/schemas/UserName"
        photo:
          type: string
          description: URL or identifier for user's profile picture
          pattern: "^.*$"
          minLength: 1
          maxLength: 255
      required:
        - id
        - name
    Error:
      type: object
      description: Standard error response object
      properties:
        code:
          type: integer
          description: Http status code 
          maximum: 999
          minimum: 0
          format: int32
        message:
          type: string
          description: Human-readable error message
          pattern: "^[a-zA-Z0-9._\\!-]+$"
          minLength: 1
          maxLength: 1000
      required:
        - code
        - message
    Path:
      type: string
      description: Path of an os
      pattern: "^(.+)\\/([^\\/]+)$"
      minLength: 4
      maxLength: 100
...
  parameters:
...
    photoId:
      name: photoId
      in: path
      description: ID of photo
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    userId:
      name: userId
      in: path
      description: ID of User
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    groupId:
      name: groupId
      in: path
      description: ID of group
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadReqError:
      description: Bad request 400
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Unauthorized 401
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Internal Server Error 500
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Not Found Error 404
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []